requirements.txt
dmgbuild>=1.6.0
pyobjc-framework-Quartz>=9.0
Pillow>=10.0.0
rumps>=0.4.0
APScheduler>=3.10.0
py2app>=0.28.0
requests>=2.31.0
PyQt6>=6.4.0
pyobjc>=9.0
desktop-notifier>=3.5.0 
regex
rubicon-objc>=0.5.0
charset-normalizer>=3.0.0
psutil>=5.9.0